+++ util/provider/provider.go
@@ -30,9 +30,9 @@ type Provider interface {
 	SetDefaultIP(ip string) error
 	AddNat(host, container string) error
 	RemoveNat(host, container string) error
-	HasShare(local, host string) bool
-	AddShare(local, host string) error
-	RemoveShare(local, host string) error
+	// HasShare(local, host string) bool
+	// AddShare(local, host string) error
+	// RemoveShare(local, host string) error
 	HasMount(mount string) bool
 	AddMount(local, host string) error
 	RemoveMount(local, host string) error
@@ -291,37 +291,37 @@ func RemoveNat(host, container string) error {
 	return p.RemoveNat(host, container)
 }
 
-func HasShare(local, host string) bool {
+// func HasShare(local, host string) bool {
 
-	p, err := fetchProvider()
-	if err != nil {
-		return false
-	}
+// 	p, err := fetchProvider()
+// 	if err != nil {
+// 		return false
+// 	}
 
-	return p.HasShare(local, host)
-}
+// 	return p.HasShare(local, host)
+// }
 
-// AddShare ...
-func AddShare(local, host string) error {
+// // AddShare ...
+// func AddShare(local, host string) error {
 
-	p, err := fetchProvider()
-	if err != nil {
-		return err
-	}
+// 	p, err := fetchProvider()
+// 	if err != nil {
+// 		return err
+// 	}
 
-	return p.AddShare(local, host)
-}
+// 	return p.AddShare(local, host)
+// }
 
-// RemoveShare ...
-func RemoveShare(local, host string) error {
+// // RemoveShare ...
+// func RemoveShare(local, host string) error {
 
-	p, err := fetchProvider()
-	if err != nil {
-		return err
-	}
+// 	p, err := fetchProvider()
+// 	if err != nil {
+// 		return err
+// 	}
 
-	return p.RemoveShare(local, host)
-}
+// 	return p.RemoveShare(local, host)
+// }
 
 // HasMount ...
 func HasMount(path string) bool {
